ci: fix bpftool install, golangci-lint Go 1.24 support, UI lint errors
- Rename scripts/onyx.service → scripts/akmon.service and
  scripts/logrotate.d/onyx → scripts/logrotate.d/akmon to match the
  Makefile SVC_NAME=akmon install paths.
- Install bpftool via linux-tools-generic (the `bpftool` package is a
  virtual package on Ubuntu 22.04 and fails to install directly).
  Symlink the concrete binary into /usr/local/bin in both ci.yml and
  release.yml.
- Bump golangci-lint from v1.61.0 to v1.64.8 so it can consume the
  go.mod go1.24.2 directive (v1.61 ships with Go 1.23 and refuses).
- AlertPanel: move Date.now() into a cutoff constant and silence
  react-hooks/purity for that line (time boundary is intentionally
  re-computed on every filter change).
- Extract buildSessions into ui/src/sessions.js; SessionGrid.jsx now
  only exports components so React Fast Refresh works.
